requirements.txt
numpy==2.2.4
openai==1.75.0
PyQt6==6.9.0
PyQt6_sip==13.10.0
python-dotenv==1.1.0
pywhispercpp==1.3.0
sounddevice==0.5.1
tavily-python 
langchain-community
langchain
wikipedia